+++ src/index.ts
@@ -10,7 +10,7 @@ export {
 export { SplatLoader } from "./SplatLoader";
 export type { SplatWorker } from "./SplatWorker";
 
-export { Splats, type SplatsOptions } from "./Splats";
+export { Splats, type SplatInput, type SplatsOptions } from "./Splats";
 export { postDecode, type SplatPostDecodeProgram } from "./postDecode";
 
 export {